backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
channels
channels-redis
daphne
django-redis
redis
django-filter
drf-spectacular
python-decouple
pytest
pytest-django
pytest-cov
pytest-asyncio
factory-boy
faker
flake8
black
python-dateutil
pytz
